Move pipeFileContents to testing file
